move usePasswordStrength file
